🆕 Price filter component
