+++ src/Homes/Filters/index.js
@@ -47,7 +47,7 @@ class Filters extends React.Component {
           />
           <Guests closeDropdown={() => this.setOpened(null)} />
           <RoomType className="hidden-xs hidden-sm hidden-md" />
-          {/* <Price text="Price" className="hidden-xs hidden-sm hidden-md" /> */}
+          <Price text="Price" className="hidden-xs hidden-sm hidden-md" />
           <Instant text="Instant book" className="hidden-xs hidden-sm hidden-md" />
           {/* <MoreFilters text="More filters" /> */}
         </Container>
